Rewrote the favorites implementation. (Apologies to Matt who knew this might happen.) This was motivated by a few things:
1. Favorites are not items, therefore a user's favorites are not an item list.
Shoehorning them into the unused item list system resulted in a lot of annoying
mismatches code that was a lot less elegant and efficient.

2. Favorites have more in common with catalog listings than stuff listings and
now they actually share code with the catalog instead of with item lists (which
will be a "My Stuff" thing if and when they do come fully into existence).

Along the way I cleaned up and unified a bunch of stuff and if it weren't for
the fact that there's a vestigial item list implementation that I left mostly
in place, this commit would have twice as many (or more) minuses as it does
plusses.