If we're going to hack the width, let's at least use a hacked with that takes up all the space we've allotted for our table.
Sigh. The "right" thing to do would be to size the table cell that contains
this panel to the desired width so that our inherited 100% size would work and
our "modular" code would not require special hackery to be used.

The really really right thing would be for HTML not to be a piece of shit and
for us to be able to say that we want the first cell in FavoritesPanel to be
just wide enough to accommodate our navigation, the second to be 15px for our
gap, the third to take up all space that's not otherwise used and the last to
be 15px for another gap. Unfortunately HTML can't actually do that, even though
it's probably the single most critical piece of functionality in any user
interface toolkit. Yay for HTML.