Removed "Sort by Remixable" because it's not actually a sort. All of the other sorts sort the entire catalog and having a sort that does not sort the entire catalog but simply sorts the results on each page is both misleading and useless.
It is not currently possible to properly sort or filter the catalog by
remixability. I am open to arguments that it is important enough that we should
make it possible, but as usual that comes at the cost of database expense and
complexity. In the meanwhile we should not pretend to have that feature and
confuse or mislead users into thinking that there are two remixable avatars in
the catalog because that's all that shows up when you "sort by remixable".