Nixed now unneeded catching of PersistenceException where it was not relevant to application logic. There are still times when we want to catch the now not quite so obvious potential failure of a database operation, and in those cases we either catch Exception or DatabaseException.
We're operating on the assumption that not having to catch PersistenceException
and log a warning a thousand times for every one time you actually want to
catch a persistence problem and act on it is a worthwhile tradeoff.

Developers will of course now forget about the fact that database operations
can fail and they'll fail to catch those persistence exceptions that should be
caught when they first write the code. That code will then fail in the field
and we'll track it down and go fix it and in theory the total pain in the ass
will be minimized. Or maybe the databases will never fail. Yeah, that's the
ticket.