It's time to play the "let's not do full table scans on the catalog database" game! Remember kids, just because it's in the database doesn't mean that you can query on it or use it in an order by clause.
Well, unfortunately you *can* and the database will dutifully do a disk based
sort on tens of millions of rows and grind everything to a halt even though
there never was and never will be a piece of software in the world that
actually *wants* that behavior. I wish there was a mode we could put Postgres
in where it would fail any query that required a full table scan to complete
properly. Then we would be reminded to create indexes for our queries when we
created them and not when the database server was mysteriously on fire.