FriendRecord is a funny little record because it's nothing but primary key. As a result, we don't want to do our normal two-phase "fetch primary key then search the cache for matching records then fetch the rest" business.
So we skip the cache entirely on these queries. That's not to say that they
won't be cached. At some point we'll implement collection caches and the whole
collection of a particular user's friends will be cached, but right now, the
individual FriendRecord instances are not cached because there's no point.

Also modified loadFriends() so that it does take advantage of cached
MemberCardRecords.