Two things:
1. I moved StuffNaviBar back into client.item and removed the unnecessary
dependency that was introduced into me.gwt.xml just to deal with having moved
StuffNaviBar out of client.item into client.stuff. Let's try to avoid digging
holes just so that we can fill them back in.

2. GWT provides an excellent feature called image bundles which we use in
various places, and were using in the previous incarnation of StuffNaviPanel.
Using an image bundle there allows the client to download one image instead of
10 when they load the Me page (the single most frequently loaded page by
registered uers). For a reason no less explicable than the first change, the
use of image bundles was thrown out when StuffNaviBar was recently changed and
replaced with the less efficient individual method. I have filled in this newly
dug hole as well.