It is a shame that the ActionScript compiler doesn't let you know that most of this code *doesn't compile*, but that's no excuse for committing it.
asdoc does helpfully let us know that it doesn't compile and this broken code
is making it difficult to build our asdocs. Instead of spending a bunch of time
fixing it, I'm going to remove it. If and when we want to reinstate it, it will
be patiently waiting in subversion.