Moved our CSS into the module public directories so that they are copied into and referenced from the sub-page directories and thus become versioned. Modules also now properly provide their CSS dependencies along with everything else.
This means editing CSS requires recompiling the page in question which is
unfortunate, but you can generally do all your CSS fiddling in Firebug and then
just record the proper values to the CSS file anyway.

I'm also looking into CssResource which is more like GWT's i18n support in that
an interface is declared with methods for every CSS stylename so that static
analysis can be applied to determine which styles are used in the code and
remove any that arent, and the styles themselves are renamed to short strings
(and other optimizations are applied to the CSS). However, I'm not yet sure
that we can use CssResource without going in and renaming a metric ton of
classes as we can't AFAIK rely on hierarchy like we do to differentiate between
ten different elements in the hierarchy all styled as Title, for example.