Holy blowtorch Batman, it's revamped chat. The salient points are:
- it makes use of the new Crowd decentralized chat services which avoid servers
subscribing to other servers' objects and just manually route all messages to
the channel host which then manually routes them back to the servers to which
the channel occupants are connected; this is basically what the old system did
through a complex network of distributed object subscriptions which were *way*
more trouble to keep up to date than they were worth

- rooms are no longer chat channels; we're back to using the room object like
civilized people; this means you can no longer click a little check box to keep
the room "channel" around for rooms you were previously occupying; that was a
feature for which the implementation complexity outweighed the delivered value
by what we would politely describe as "a great deal"; the main reason that
quixotic journey was set upon was to allow players to keep their most recent
room's chat open while they were in a game; we can accomplish this in a much
simpler way in our new simpler world

- there are no longer occupant lists for group channels (rooms still have
occupant lists which actually display the occupants of the room, which you
might think was what they previously displayed, but you would be wrong); at
some point I'll add a /cwho command that will request the names of everyone in
the current channel; since channels are decentralized it is no longer possible
to broadcast to every member of the channel the fascinating comings and goings
of everyone else in the channel

- group channels now automatically activate for every online member of a group
any time someone in that group speaks on the group channel (they don't pop to
front, you just get a glowing tab in your header); it is not yet possible to
say "don't show me this tab for the rest of this session unless I open it
myself"); we'll add that capability at some point in the future.

I still need to add/remove people to/from channels when they join/leave groups.
There are no doubt other bugs and weirdness as well. This chat stuff was angora
rabbit hairy before I came in with the blowtorch and now we've got it down to
just alpaca hairy or thereabouts.