Cleaned up and documented the chat message expiry process. Messages from a non-visible tab are properly displayed when you show the previously hidden tab without weird hackery in displayMessage.
We no longer restore the last minute's worth of chat when reopening a friend or
jabber tab. If you have unexpired messages in the tab, they'll be reshown,
which gives you the last 15 seconds, but I don't see why we should do any
backbending and complexifying to give you a longer window for those channels
when we don't for any others and that's plenty of time to catch the "I said bye
and closed the tab and then you said one more thing and I reopened the tab and
saw a bit of context."

If you want more than that, turn on your chat history and it'll all be there.