Sorry for losing my cool today, but dayamn!
4 days of discussion, a wiki page....
OR: this code. Done.

Just track a blingAffiliate separately. It's keyed off the same aff= tag
initially. We do need to change the way we do embeds so that we put that
tag directly on the SWF url. That's OK because we end up redirecting to the
current versioned swf file anyway. Easy peasy.
- TODO: undo the cOmPlExITY of today's affiliate-map-updating and just use
  MemberRecord's new blingAffiliate to do it in like one line.
- Go smoke a fat one, because this is how easy shit should be.

Please do not adjust your seatbelts. The existing TrackingCookie,
whatever-the-hell, still works however-it-does.