The measure of a programmer is not how they handle success, but how they handle failure.
Truly, it is a testament to the myriad achievements of mankind that we can
carelessly marshal the resources of dozens of routers across huge spans of the
globe to transmit blank or invalid email addresses to our servers wherein we
can consume expensive electricity and generate heat in our colo to determine
that said email is blank or invalid, expend more heat and energy logging a
stack trace to our server logs which will cause (even more expensive) human
beings to spend time and energy scanning past those spurious errors in the
logs.

But the wise programmer checks the validity of the email addresses on the
client and pats himself on the back for doing his part in the fight against
global warming and the fight for user interfaces that don't silently fail.

The wise programmer also does not send a TextArea as an argument in a
CommandEvent. What sort of cracksmoking was going on there?