Next step in the new entry vector tracking implementation. In this installment:
- We're now tracking the "visitor id" in a separate way, with a cookie that gets cleared out 
upon login, and a new "visitor id" field on the member record. The old data from 
ReferralRecord gets migrated over, and we'll be able to delete it at some convenient time in 
the future. 

- Started converting over from the old URL affiliate string, to the new uber-entry-vector 
format, which gets logged to Panopticon, but no longer to the user database. Panopticon should 
already reconcile them.

The old system is still active, just getting worked around.