Changed initialization a bit so that we disallow doing any database fooling around during the Guice dependency resolution phase and we also require that all dependency resolution that requires repositories be done at this time and not lazily later.
We already had the notion of "construct" and "init" for our managers which
meant that during dependency resolution everyone just got contsructed and no
initialization took place. Now we have the same thing for our repositories.

This ensures that before the server starts doing anything important, we are
sure that our database house is fully in order, all schema migrations are run
and everything is kosher. It also ensures that no one tries to jump the gun and
start migrating schemas or reading or writing data before we're ready.

I also removed the ability to configure your server to lazily initialize
repositories. It takes less than a second to just initialize them all, so let's
just save ourselves all the trouble and make sure that all developers
immediately run schema migrations just like we do on dev and production.