Once upon a time the UserAction system into being for the purpose of tracking what players did for periodic humanity assessment. Then people came along and forced poor UserAction to play a part in recording flow expenditure for user behavior analysis. Then more people came along and forced poor UserAction into playing a part in reporting the details of financial transactions to users in their transaction history. By this time, UserAction was abused and confused and all the code that he was involved in was a giant mess.
Fortunately, UserAction gave some thought to his role in life and sat down and
negotiated with all of the services that he was working for and came up with a
plan where they would demand less of him and he would provide more and in a
more predictable way. He was happy, his employers were happy, and everyone
lived happily ever after.