We now have DataMigration, so please no longer use SchemaMigration to do your data migration. It is prone to error. Data migrations will be run after all schema migrations have completed.
This means you don't have to worry about columns not being added before your
data migration tries to populate them or something crazy like that. However,
you do have to worry about dropped columns. Say you wanted to move a column
from one record to another, you should add the new column to the new table, and
add a data migration that copies the data from the old table to the new.

Then after sufficient time has elapsed that everyone that you care about is
known to have run the migration (certainly the production servers are included
in that group), you can go in and add a SchemaMigration.Drop to rmeove the old
column and *remove* the data migration from the code.

This means that anyone that did not run the migration during the available
window simply does not get it. Those are the breaks. You have to remove the
column from the PersistentRecord class at the time that you add the Drop, so
your data migration is no longer going to compile anyway.

Schema and data migration are complicated and these compromises helps to keep
things relatively simple.