AVRG payouts were something of a pig's breakfast. Now they behave in a manner approximating the way normal game payouts do.
We accumulate time for a player, and when they complete a task, we note their
accumulated playtime as well as their accumulated coins and we reset the timer
for the next task. When the player leaves *or* the game is shutdown, we pay out
all unpaid out players. Only when a player is persistently paid out do we
potentially recalculate our payout factor.

Currently we're still recording quest completion to the database which I
believe is only being done because the quest system was coopted for AVRG payout
factor calculation and when we ditched the quest system in favor of tasks, we
had to leave this cruft in there to preserve payout recalculation. When I
confirm this to be true, I'll rip that out as well and make AVRGs use the same
system that normal games use and all will become well and right with the
universe.