Let there be artist freakoutery, but we cannot just chop the preview size in half and magically expect things to work.
HTML is awesome and therefore does not provide a mechanism for saying "please
scale this image to fit into this space, or not at all if it is smaller than
the space", so we have to do a bunch of major fucking around to ensure that we
can properly scale images by their width or height or not at all where
appropriate.

That involves computing the image's constraint in a certain dimension (width or
height) and using that when the time comes to scale. Preview images have their
constraint calculated for the preview size and if you take a media descriptor
for such an image and tell MediaUtil to display it in PREVIEW_SIZE/2 it will
fail horribly for images whose original size is between PREVIEW_SIZE/2 and
PREVIEW_SIZE as those images will be displayed totally unscaled (and hence way
too large) and will fuck up the layout.

I hate the web. We shouldn't even have this problem.