Goddamned terminology. Not only do we have to create the Brave New Whirled, we have to create non-confusing ways of talking about it. Catalog original, catalog prototype and catalog clone are not cutting the mustard.
Let's try:

Listed original - an original item that has been listed in the catalog and now
serves as the working copy from which new catalog listings are made.

Catalog master - a snapshot of an item that is listed in the catalog and which
serves as the source for all purchased copies of that catalog listing.

Catalog clone - a clone of the catalog master. What you have if you buy a
catalog listing.

This also fixes a bug with getSuiteId().