GameDetailRecord was not being properly updated when a game was listed because catalogId was not being filled into the master record by ItemRepository.insertListing and as a result isCatalogMaster() was failing.
GameDetailRecord was also not being properly updated when a game was delisted
because we were not notifying ItemLogic when an item was deleted because its
catalog master record was deleted, only when an item was deleted per request
from a user. We were also not actually clearing out the listedItemId in
GameDetailRecord because we were operating under the erronious assumption that
catalog master would never be deleted, which was doubly erroneous because even
if that were the case we still want to clear out a game's listedItemId if the
listing is removed even if the master is kept around because people purchased
copies of it. Now we properly notify ItemLogic that an item was deleted if a
catalog master is deleted and we unlink a game from its catalog listing when
the listing is removed regardless of whether or not that resulted in the
catalog master being deleted.