Because Photo was claiming that photoMedia was is default furniMedia, it would fail to store its furniMedia when photoMedia that was small enough to be the furniMedia without scaling was uploaded. Then it would end up with null furniMedia which would cause the thumbnail media to be used as the preview. No good. We no longer claim that our photoMedia is our default furniMedia.
getPreviewMedia() should also never return the photoMedia as that will not have
had its constraint computed at the right size and is also probably enormous. It
should use the furniMedia if we have one and if not use getThumbnailMedia() to
use either the thumbnail media or the default thumbnail media as appropriate.