Let's start tracking when players and guests start their GWT sessions, so that we can compare Flash logins vs. GWT logins.
Tracking the beginning of a 'guest' GWT session is a little difficult, though, because a guest 
doesn't log on per se - rather, they /fail/ to log on, and get immediately logged off. 
But we've overloaded the term 'logoff' to cover a number of semi-related conditions: 
- player was logged on, and they clicked the 'logoff' button
- player attempted to log on and failed unexpectedly
- random guest landed without any credentials

For metrics, we care about case #3 separately from the other two. So let's make the various 
types of 'logoffs' explicit, and record them along with registered players' logons.

WRLD-386