Oh I'm being naughty today. I want links to "Level Packs" and "Item Packs" on the Stuff page but the navigation mechanism there, which I never liked, isn't scalable and can't support it. So I did what I wanted to do from the beginning which was to use the shop-style sidebar navigation which is both scalable and allows what you care about, your stuff, to be higher on the page.
The only argument for not doing this was a concern that the stuff page look too
much like the shop page. There are already a bunch of differences and in the
abstract this argument is dubious because we use the same top-level navigation
and sub-navigation on all of our pages, so obviously we don't care *that* much
about pages looking similar. I think there are already plenty of visual
differntiators and if we feel that we want more then we can find other ways
than the navigation to differentiate.