Minor updates (that don't seem to really matter) to match the new SDK.
Strangely, the new unloadAndStop() isn't even present on Loader objects
in the new fp10. The new version of the SDK I checked-in yesterday doesn't
even have 10 support yet, so I'll try screwing with a nightly build to
see if I can get it to work. That's truly weird though, it would mean
that the Loader class is compiled into the SWF? That *can't* be!