A couple more things while I'm micromanaging. One style note:
  Predicates.<SceneRecord> alwaysTrue()

looks like a type declaration with the thing on the left a generic type and the
thing on the right a declared variable or argument.

We want to distinguish visually what is actually going on which is an explicit
generic type instantion so we write:

  Predicates.<SceneRecord>alwaysTrue()

Further, this explicit generic type instantion is kind of like a type cast
which meshes nicely with my preferred formatting for a type cast:

  (SomeType)someVar

and reinforces my dislike for this style (practiced by some engineers who will
remain nameless):

  (SomeType) someVar

As I mentioned to Bruno in a length and pedantic argument, there is an English
language linguistic precedent for unary prefix operators (! -- etc.) to be
cuddled up next to their operand (--foo !foo etc.) rather than be separated by
a space. This contrasts with binary infix operators which (English language)
linguistics suggests should be surrounded by space.

In my mind, a cast is a unary prefix operator.