Another totally minor drive-by optimization. Avoid re-loading the in-progress badges in a place where we've just loaded them.
Ok: I've looked through this and I have a plan for handling the figure-out of
whether we need to compute new inprogress badges. I was going to fallback to
adding a HIDDEN2 badgeType, but that's so ugly and it turns out it won't work
without some other rejiggering.

So, what I'm going to do is have a version number in badgeType. (I could
autocompute a hashcode, but then it will change when we remove as well as
add badges...). Everywhere where the badges might need to be recalculated is
called from a place where we've got either the member record or object, so
once we've computed the badge types for a user we set their BadgeVersion to
the value in BadgeType and then we never need to recompute unless we up the
badge version.