Usually a disconnected player has a period of time to reconnect and automatically resume their play-session. However, for usability reasons we've decided that a client should not auto-load AVRG media, e.g. after a page reload, without the player deliberately restarting the AVRG.
But we also really don't want a prolonged situation where the server-side usercode treats the player as playing the AVRG and receives e.g. movement events for them, while the client doesn't have the client-side media loaded. The inescapable conclusion would seem to be that AVRG sessions need to be terminated server-side on any disconnect.

The reason this is not necessarily a big problem is that AVRGs have to be written to be very reentrant anyway; people drop in and out of them all the time and expect all but the most ephemeral state to be persisted.

Counter-suggestions happily accepted.