External media! Built-in youtube support! - Changed the plan on external media, since we usually want to know what   we're dealing with from the mimeType, so each type of external media   (just the 1 right now) will have its own mimeType. - So that metadata doesn't need to be in the file. Instead of encoding as   XML, I decided to just use "URLVariables" format, because URLLoader   can load and return an Object with all the goodies parsed. It's shorter   and accomplishes the same thing. I also omitted a version number, but   each kind of external media can do what it wants with that.
This is a first-pass, there are some bits that need to change before we can
ship; I'll get right on those bits.
After that, there is some cleanup that should be done.
After that, I'll get on the new snazzy UI for videos, because all this is
still using the "SimpleVideoDisplay" class.