At Zell's request, made following much more robust in the face of spastic leaders who blaze from server to server with no consideration for their poor followers who are struggling to keep up.
Now we dispatch follow-the-leader notifications via MemberNodeActions which may
still also fail if the member in question is currently in the middle of
switching servers (likely), but now if someone misses a particular follow
notification they aren't dropped like a hot potato. They'll very likely get the
next follow notification and be back with the in-crowd lickety split.