Tracked down the final problems with seeking...
Hot damn! It works! If this isn't v1.0 of our video player, it's
at least v.99, because there are a few inconsistencies with the
way FLV videos and Youtube videos work.
(Those are:
  - what is displayed at the end
  - how immediately the seek takes effect
  - what happens to playback if you seek while playing or paused
)

Oh, and since I didn't want to re-implement the volume popup
without flex, the volume button simply toggles the video between
full volume and muted.

I still need to:
- clean up logging
- convert legacy youtube videos.
- test out starting the youtube player differently, possibly
  fixing the flakiness that sometimes happens.