A useful thing to consider when writing a database query is: "Am I playing nicely with the cache?"
In the case of testing for greetership, we will quite probably have a member's
record in the cache and can just grab that and check it for greetership. If we
don't have it cached and cache it as a result of our test, any of ten dozen
other places in the code that load up a member's record will benefit from our
caching.

Doing a collection query just to read a flag out of MemberRecord.flags will
result in caching an enormous key and a tiny value and no other code in the
rest of the codebase will benefit from that use of the cache.