A framework for structured stats collection and reporting.
The /status servlet exists so that we can debug strange runtime behavior via a
mobile phone. For more voluminous and detailed runtime statistics, we want the
ability to sort and summarize the data that comes from a bunch of nodes and
that's way more than the /status servlet can and should handle.

This framework allows us to collect data from multiple servers and combine it
into a StatsModel in whatever way makes sense for the stats in question and
then we'll have an admin page that displays these stats using a fancy GWT table
that slices, dices and even make julienne fries.