Avatars were showing up without memories. I believe this happened because we used bodyEntered() to store the avatar's memories in the RoomObject. But that's too late: the client will have already begun loading entity media by the time the memories arrive. If the timing is just so, the avatar media could surely initialize *before* they arrive. This would lead to a memory-less avatar.
My initial urge was to send the memories earlier; I tried hijacking insertOccupantInfo() for this purpose, but then memories get put into the next room *before* they get taken out of the old room, and that's a whole lot worse.

So here's my solution. I hope it's not considered too gauche to write logic into the MemberObject like this, but it's the single place where we know we're being notified at precisely the right moment. I've run around on my test machine quite a bit and nothing's gone awry yet.