The true motivation for the timeless way of updating is revealed: MSOY was just blithely totally recreating MemberInfo whenever anything changed which is Not The Right Thing (tm).
The Right Thing (tm) is to obtain the canonical MemberInfo record, update it
with just the things that have changed, and dispatch that as an update. Now it
is only possible to do things The Right Way (tm) and so we do so.

Currently the RoomManager.updateOccupantInfo() override is commented out
because this was further wrongness on top of our wrong handling of OccupantInfo
updates both of which were no doubt the source of the bugs involving this whole
business. I need to look into where else stone tofu are assigned or not
assigned and make sure we're doing the right thing.