For great brevity!
Not yet commented or in global use, I thought I'd give mdb a chance
to give feedback on the concept and the method names.

Basically, in most places where you previously had to import a particular adapter
you can just do:
_someService.makeRequest(_ctx.getClient(), (... args), _ctx.resultListener(resultHandlerFn));
Errors are logged and go to chat. There are a bunch of other options for
customizing these things, which I'll put in the TBD method comments.