Enable ping_debug on the dev server to further investigate pesky game server death. I think what is happening is that the game server is starting up while something else is grinding le crap out of dev1 and it is prevented in doing its initial handshakes with the server for 90 seconds which causes it to stick a fork in itself before the world server ever even knew it existed, and the world server then sits around forever waiting for it to come back.
We can of course listen to the process forekd off by the world server to hear
whether it died and check to see if we ever heard from that game server and
start one up anew in that case but it's annoying to have to write a bunch more
code and complexify things even more just because sometimes someone is grinding
the shit out of the dev1 server when the autoreboot happens.

I'd rather turn off the autoreboot or stop grinding the dev server.