More FBConnect wrangling. We need things at the top-level again so that we can post stories to a person's feed. Doing this from the server turns out to be more complicated (requiring us to communicate and maintain the FB session information on the server) and it prevents us from posting things to the feeed of someone who did not connect to FB during this session (or ever).
By posting from the client, we circumvent all that PITA, and we get the added
bonus that even if someone has never done any sort of Facebook connectery, if
they have a Facebook account and are logged into it, they'll get a popup asking
if Whirled can post to their feed anyway (which is sort of like what Netflix
and co. do with Facebook; it's a twisty maze of APIs all alike).

This does mean we're going to have to route feed postings through the Flash
client, but hopefully we can streamline that process to the point where it's
minimally ugly and annoying.