One more in a seemingly endless stream of avatar memory resolution race conditions fixified. This time, switching avatars while in a room would kick off the database-hitting memory-reading invoker unit only when the occupant info change was detected.
That's too late. We have to get this data into place before we tell the client what we're doing, before the media starts loading.

So now we read the memories in MemberManager.finishSetAvatar(), put them into MemberLocal the same way that normal room-to-room movement works, and manually transfer the memories to whatever room the user is in.