Bugfix: entities now "snoop" chat, so that any entities you control can react to chat prior to it being filtered (which includes muting). I was going to just hack this together (woulda been so nice) by just reacting to the raw dobj message, but there's no guarantee that we'll process it prior to the ChatDirector, and the ChatDirector modifies the message during processing. I suppose I could have changed that, but that seemed a little too cart-before-horse.
I created the ChatSnooper interface, and use that here, but
unfortunately we still are a ChatDisplay as well, because
an avatar's reaction to speaking should only occur if
it actually spoke. Unfortunately that means that when
someone talks we look up their sprite first to notify
room entities, and then we look it right back up later
to notify it that it talked. If you had that person
muted, the second part wouldn't happen.