Media uploading/editing: - Moved the button to take a snapshot with your webcam out to the top-level HTML, where it's more visible.
Still need to finish full media upload revamp, to
rethink ImageChooserPopup.

Also, this is causing me annoyance:
- lining up flash and html buttons
- The button skin needs to be embedded in this new wee flash applet,
in addition to being embedded in another right next to it.
Probably I should just make the whole buttony area next to the
preview a flash applet, and have it do all the buttons.
But, this is still progress, so I'll check this in for now.