Omigod, it all seems to work. - Detect when we're in an embed/stub running with an old version of flash, and prompt the user to upgrade. We do this by simply redirecting them to whirled.com for the same room/game they're viewing, and then the GWT code to update their flash player will kick in. After the upgrade, they should be back at whirled.
- Detect when we're in a stub, the user has the right version of flash, but
the stub is holding us back. Display a message indicating that the page
owner needs to update their stub, and provide a link to redirect to the
same room/game at whirled.com.
When we're in this situation, it turns out that there's no way to prevent
the faceplant that flex does on startup, but luckily it does start the
preloader first and that's where this code is. If you don't have the debug
player, you'll never know that flex faceplanted.

These messages need to be prettied-up, possibly by loading in a snazzy
image to accompany them and actually make the user curious about what can
be seen by clicking the link. Also, there's no i18n here. We're in the
preloader for fuck's sake, and I'm already dragging way more classes in
that I'd like to.
(Added: DeploymentConfig (small), MsoyParameters which is small but
brings in ParameterUtil, which brings in Log, which brings in StringUtil.
Crap.)