PersistingUnit handles a lot of the boilerplate of doing database stuff with an invocation listener. This also allows us to throw invocation exceptions instead of setting a failure code and returning and then reporting failure in handleSuccess(). (We have successfully failed! :)
Nixed calls to user.isActive() as it's a little dangerous to call any dobj
method from the invoker thread (in this case the worst that would likely happen
is isActive() returns true when it should return false, but it sets a bad
example; we allow the call to MemberObject.getMemberId() as that should be
annotated @AnyThread).

Also nixed a duplication of the _avatar.ownerId argument in a warning log.