Game shop revamp. It consists of a couple of things:
1. Instead of displaying a bunch of potentially empty categories, we just
display everything in every category in one (not very) giant list. This is
egregiously expensive (potentially requiring 12 expensive databsae queries). I
am predicating this change on the following math:

probability_of_paying_us_money(someone_viewing_game_shop) == high
number_of_viewers_of_game_shop == low

If the former assumption turns out to be false then we might as well scrap the
game shop entirely. If the latter assumption turns out to be false, then we'll
take a break from sipping pina coladas on our pile of money long enough to add
the engineering complexity needed to make this more efficient.

2. The Games section of the shop no longer links to the game item's detail
page. It links to the Game Shop for that game. In the Game Shop now lives the
game item, clicking that will allow you to buy the game item. For games that
have nothing in their shop, this means it's one extra slightly weird step to
buy the game. Every game should have something in its shop. A game that is
failing to merchandise is failing to take advantage of Whirled. We should focus
our energies on solving that problem for games that have it.

If this new game shop world order works more nicely, then we can invest more
time into it later to make the display sexier (specifically to list the
different types of items under separate headings to make it a big clearer
what's going on for games that have a lot of different kinds of merchandise).