There are three big problems with the way we're tracking client actions, the main upshot being that it just doesn't work. Both problems stem largely from the following sequence:
- user clicks button
- click listener to take user to new page is fired, browser queues up page change
- click listener to report user action is fired, browser queues up RPC request
- browser executes page change, cancels all outstanding RPC requests

So the RPC request to report the user action never goes through (or at the very
least it's a race condition; it definitely never goes through in the fast
environment on a developer workstation).

This problem is solved in this commit wherein the user action is delivered to
the top-level frame which handles submitting it to the server. The top-level
frame never goes away.

The second problem stems from the fact that the sequence shown for the above
problem is a simplification. In reality, if the page change click listener
fires before the action report click listener, the action report still
sometimes fails with weird JavaScript exceptions. So what really needs to
happen is that the action report click listener needs to be fired before the
page change listener. This is a PITA because naturally one wants to write:

   Button button = new Button("Go!", Link.createListener(Page.FOO, bar));
   MsoyUI.addTrackingListener(button, "Went!", "somedata");

but that results in things being wired up in the wrong order. We can't have
addTrackingListener() magically put itself first (no API support), nor can we
have it remove any registered click listeners and add itself and then
reregister the other listeners because there's no API for enumerating
registered click listeners. So we're going to have to change every single place
that uses addTrackingListener() and addTestTrackingListener() to do things
specially. Yay! That will come in a subsequent commit.

The third problem is a big one and I'm not yet sure how we're going to solve
it. Sometimes we do things like this:

    Hyperlink link = Link.create("Go!", Page.FOO, bar);
    MsoyUI.addTrackingListener(link, "Went!", "somedata");

In this case, there's no page change listener. The link is a link and when it
is clicked the browser unloads the current page, stops all executing JavaScript
and places a giant boot to the ass of the action report click listener. I don't
think there's anything we can do in this case except to either put the action
tracking on the destination page or not use a hyperlink (which is annoying
because we try to use hyperlinks everywhere we can because that makes Whirled a
usable website rather than a giant JavaScript hack).

The moral of the story? The web sucks as an application development platform.
The curse I leveled against Brendan Eich ("inventor" of JavaScript) ten years
ago is up for renewal and this time I'm going to pay for the 50 year discounted
rate.